separate dtos and interfaces
